requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
discord-typings==0.6.0
emoji==2.6.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.3
idna==3.4
interactions.py==5.7.0
multidict==6.0.4
python-dotenv==1.0.0
sniffio==1.3.0
tomli==2.0.1
typing_extensions==4.7.1
yarl==1.9.2
